Backend/requirements.txt
Flask
flask-cors
pyttsx3
google-generativeai
